Update README.md
A tech community builder through code and collaboration, I'm a software engineer who's not just passionate about programming but equally committed to fostering and strengthening the bonds that fuel innovation. My mission extends beyond lines of code, aiming to unite and empower tech enthusiasts to create a dynamic and supportive ecosystem for all.